requirements.txt
absl-py==2.2.2
apturl==0.5.2
astunparse==1.6.3
blinker==1.4
Brlapi==0.7.0
cachetools==5.5.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.40.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
httplib2==0.14.0
idna==2.8
importlib-metadata==8.5.0
jinja2==3.1.6
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.7
MarkupSafe==2.1.5
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.4.0
packaging==25.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.9
pyRFC3339==1.1
PySide6==6.2.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
rsa==4.9.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
shiboken6==6.2.4
simplejson==3.16.0
six==1.14.0
systemd-python==234
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
typing-extensions==4.13.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usd-core==25.5
wadllib==1.3.3
werkzeug==3.0.6
wrapt==1.17.2
xkit==0.0.0
zipp==3.20.2
